dev to main

+++ frontend/src/api/axios/axios.custom.ts
@@ -1,9 +1,9 @@
+import { ClubUserDto } from "@/types/dto/lent.dto";
 import CabinetStatus from "@/types/enum/cabinet.status.enum";
 import CabinetType from "@/types/enum/cabinet.type.enum";
-import instance from "./axios.instance";
+import instance from "@/api/axios/axios.instance";
 
-const axiosLogoutUrl = "/auth/logout";
-
+const axiosLogoutUrl = "/v4/auth/logout";
 export const axiosLogout = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosLogoutUrl);
@@ -13,7 +13,7 @@ export const axiosLogout = async (): Promise<any> => {
   }
 };
 
-const axiosMyInfoURL = "/api/my_info";
+const axiosMyInfoURL = "/v4/users/me";
 export const axiosMyInfo = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosMyInfoURL);
@@ -23,57 +23,25 @@ export const axiosMyInfo = async (): Promise<any> => {
   }
 };
 
-const axiosUpdateCabinetMemoURL = "/api/lent/update_cabinet_memo";
-export const axiosUpdateCabinetMemo = async (
-  cabinet_memo: object
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      axiosUpdateCabinetMemoURL,
-      cabinet_memo
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosUpdateCabinetTitleURL = "/api/lent/update_cabinet_title";
-export const axiosUpdateCabinetTitle = async (
-  cabinet_title: object
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      axiosUpdateCabinetTitleURL,
-      cabinet_title
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
 // V3 API
-// TODO
-// 차후 API 확인 후에 URL 수정
-const axiosLocationFloorURL = "/api/cabinet_info";
-export const axiosLocationFloor = async (): Promise<any> => {
+const axiosBuildingFloorURL = "/v4/cabinets/buildings/floors";
+export const axiosBuildingFloor = async (): Promise<any> => {
   try {
-    const response = await instance.get(axiosLocationFloorURL);
+    const response = await instance.get(axiosBuildingFloorURL);
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosCabinetByLocationFloorURL = "/api/cabinet_info/";
-export const axiosCabinetByLocationFloor = async (
-  location: string,
+const axiosCabinetByBuildingFloorURL = "/v4/cabinets/buildings/";
+export const axiosCabinetByBuildingFloor = async (
+  Building: string,
   floor: number
 ): Promise<any> => {
   try {
     const response = await instance.get(
-      `${axiosCabinetByLocationFloorURL}${location}/${floor}`
+      `${axiosCabinetByBuildingFloorURL}${Building}/floors/${floor}`
     );
     return response;
   } catch (error) {
@@ -81,8 +49,11 @@ export const axiosCabinetByLocationFloor = async (
   }
 };
 
-const axiosCabinetByIdURL = "/api/cabinet_info/";
-export const axiosCabinetById = async (cabinetId: number): Promise<any> => {
+const axiosCabinetByIdURL = "/v4/cabinets/";
+export const axiosCabinetById = async (
+  cabinetId: number | null
+): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.get(`${axiosCabinetByIdURL}${cabinetId}`);
     return response;
@@ -91,8 +62,9 @@ export const axiosCabinetById = async (cabinetId: number): Promise<any> => {
   }
 };
 
-const axiosLentIdURL = "/api/lent/";
-export const axiosLentId = async (cabinetId: number): Promise<any> => {
+const axiosLentIdURL = "/v4/lent/cabinets/";
+export const axiosLentId = async (cabinetId: number | null): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.post(`${axiosLentIdURL}${cabinetId}`);
     return response;
@@ -101,7 +73,7 @@ export const axiosLentId = async (cabinetId: number): Promise<any> => {
   }
 };
 
-const axiosMyLentInfoURL = "/api/my_lent_info";
+const axiosMyLentInfoURL = "/v4/lent/me";
 export const axiosMyLentInfo = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosMyLentInfoURL);
@@ -111,32 +83,48 @@ export const axiosMyLentInfo = async (): Promise<any> => {
   }
 };
 
-const axiosReturnURL = "/api/lent/return";
-export const axiosReturn = async (): Promise<any> => {
+const axiosUpdateMyCabinetInfoURL = "/v4/lent/me/cabinet";
+export const axiosUpdateMyCabinetInfo = async (
+  title: string | null,
+  memo: string | null
+): Promise<any> => {
   try {
-    const response = await instance.delete(axiosReturnURL);
+    const response = await instance.patch(axiosUpdateMyCabinetInfoURL, {
+      title,
+      memo,
+    });
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosSendCabinetPasswordURL = "/api/lent/return-memo";
+const axiosReturnURL = "/v4/lent/return";
+export const axiosReturn = async (): Promise<any> => {
+  try {
+    const response = await instance.patch(axiosReturnURL);
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosSendCabinetPasswordURL = "/v4/lent/return-memo";
 export const axiosSendCabinetPassword = async (password: string) => {
   try {
     const response = await instance.patch(axiosSendCabinetPasswordURL, {
-      cabinet_memo: password,
+      cabinetMemo: password,
     });
   } catch (error) {
     throw error;
   }
 };
 
-const axiosMyLentLogURL = "/api/my_lent_info/log";
+const axiosMyLentLogURL = "/v4/lent/me/histories";
 export const axiosMyLentLog = async (page: number): Promise<any> => {
   try {
     const response = await instance.get(axiosMyLentLogURL, {
-      params: { page: page, length: 10 },
+      params: { page: page, size: 10 },
     });
     return response;
   } catch (error) {
@@ -145,7 +133,7 @@ export const axiosMyLentLog = async (page: number): Promise<any> => {
 };
 
 // Admin API
-const axiosAdminAuthLoginURL = "/api/admin/auth/login";
+const axiosAdminAuthLoginURL = "/v4/admin/auth/login";
 export const axiosAdminAuthLogin = async (
   id: string,
   password: string
@@ -161,10 +149,11 @@ export const axiosAdminAuthLogin = async (
   }
 };
 
-const axiosAdminCabinetInfoByIdURL = "/api/admin/cabinet/";
+const axiosAdminCabinetInfoByIdURL = "/v4/admin/cabinets/";
 export const axiosAdminCabinetInfoByCabinetId = async (
-  cabinetId: number
+  cabinetId: number | null
 ): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.get(
       axiosAdminCabinetInfoByIdURL + cabinetId
@@ -175,43 +164,23 @@ export const axiosAdminCabinetInfoByCabinetId = async (
   }
 };
 
-const axiosGetCabinetStateURL = "/api/admin/cabinet/count/floor";
-export const axiosGetCabinetState = async (): Promise<any> => {
-  try {
-    const response = await instance.get(axiosGetCabinetStateURL);
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosAdminReturnURL = "/api/admin/return/cabinet/";
-export const axiosAdminReturn = async (cabinetId: number): Promise<any> => {
-  try {
-    const response = await instance.delete(axiosAdminReturnURL + cabinetId);
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosReturnByUserIdURL = "/api/admin/return/user/";
+const axiosReturnByUserIdURL = "/v4/admin/return-users/";
 export const axiosReturnByUserId = async (userId: number): Promise<any> => {
   try {
-    const response = await instance.delete(axiosReturnByUserIdURL + userId);
+    const response = await instance.patch(axiosReturnByUserIdURL + userId);
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosBundleReturnURL = "/api/admin/return/bundle/cabinet";
+const axiosBundleReturnURL = "/v4/admin/return-cabinets";
 export const axiosBundleReturn = async (
   cabinetIdList: number[]
 ): Promise<any> => {
   try {
-    const response = await instance.delete(axiosBundleReturnURL, {
-      data: cabinetIdList,
+    const response = await instance.patch(axiosBundleReturnURL, {
+      cabinetIds: cabinetIdList,
     });
     return response;
   } catch (error) {
@@ -219,73 +188,29 @@ export const axiosBundleReturn = async (
   }
 };
 
-const axiosUpdateCabinetTypeURL = "/api/admin/cabinet/lentType/";
-export const axiosUpdateCabinetType = async (
-  cabinetId: number,
-  lentType: CabinetType
-) => {
-  try {
-    const response = await instance.patch(
-      `${axiosUpdateCabinetTypeURL}${cabinetId}/${lentType}`
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosBundleUpdateCabinetTypeURL = "/api/admin/cabinet/bundle/lentType/";
-export const axiosBundleUpdateCabinetType = async (
-  cabinetIdList: number[],
-  cabinetType: CabinetType
+const axiosUpdateCabinetsURL = "/v4/admin/cabinets/";
+export const axiosUpdateCabinets = async (
+  cabinetIds: number[],
+  lentType: CabinetType | null,
+  status: CabinetStatus | null
 ): Promise<any> => {
   try {
-    const response = await instance.patch(
-      `${axiosBundleUpdateCabinetTypeURL}${cabinetType}`,
-      cabinetIdList
-    );
+    const response = await instance.patch(axiosUpdateCabinetsURL, {
+      cabinetIds,
+      lentType,
+      status,
+    });
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosUpdateCabinetStatusURL = "/api/admin/cabinet/status/";
-export const axiosUpdateCabinetStatus = async (
-  cabinetId: number,
-  status: CabinetStatus
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      `${axiosUpdateCabinetStatusURL}${cabinetId}/${status}`
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosBundleUpdateCabinetStatusURL = "/api/admin/cabinet/bundle/status/";
-export const axiosBundleUpdateCabinetStatus = async (
-  cabinetIdList: number[],
-  cabinetStatus: CabinetStatus
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      `${axiosBundleUpdateCabinetStatusURL}${cabinetStatus}`,
-      cabinetIdList
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosSearchByIntraIdURL = "/api/admin/search/intraId/";
+const axiosSearchByIntraIdURL = "/v4/admin/search/users-simple";
 export const axiosSearchByIntraId = async (intraId: string) => {
   try {
-    const response = await instance.get(axiosSearchByIntraIdURL + intraId, {
-      params: { page: 0, length: 10 },
+    const response = await instance.get(axiosSearchByIntraIdURL, {
+      params: { name: intraId, page: 0, size: 10 },
     });
     return response;
   } catch (error) {
@@ -293,14 +218,11 @@ export const axiosSearchByIntraId = async (intraId: string) => {
   }
 };
 
-const axiosSearchByCabinetNumURL = "/api/admin/search/cabinet/visibleNum/";
-export const axiosSearchByCabinetNum = async (
-  number: number,
-  floor?: number
-) => {
+const axiosSearchByCabinetNumURL = "/v4/admin/search/cabinets";
+export const axiosSearchByCabinetNum = async (number: number) => {
   try {
-    const response = await instance.get(axiosSearchByCabinetNumURL + number, {
-      params: { floor: floor },
+    const response = await instance.get(axiosSearchByCabinetNumURL, {
+      params: { visibleNum: number },
     });
     return response;
   } catch (error) {
@@ -308,29 +230,39 @@ export const axiosSearchByCabinetNum = async (
   }
 };
 
-const axiosSearchDetailByIntraIdURL = "/api/admin/search/";
+const axiosSearchByCabinetNumSimpleURL = "/v4/admin/search/cabinets-simple";
+export const axiosSearchByCabinetNumSimple = async (number: number) => {
+  try {
+    const response = await instance.get(axiosSearchByCabinetNumSimpleURL, {
+      params: { visibleNum: number },
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosSearchDetailByIntraIdURL = "/v4/admin/search/users";
 export const axiosSearchDetailByIntraId = async (
   intraId: string,
   page: number
 ) => {
   try {
-    const response = await instance.get(
-      axiosSearchDetailByIntraIdURL + intraId,
-      {
-        params: { page: page, length: 10 },
-      }
-    );
+    const response = await instance.get(axiosSearchDetailByIntraIdURL, {
+      params: { name: intraId, page: page, size: 10 },
+    });
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosDeleteCurrentBanLogURL = "/api/admin/log/ban/";
-export const axiosDeleteCurrentBanLog = async (userId: number) => {
+const axiosDeleteCurrentBanLogURL = "/v4/admin/users/";
+export const axiosDeleteCurrentBanLog = async (userId: number | null) => {
+  if (userId === null) return;
   try {
     const response = await instance.delete(
-      axiosDeleteCurrentBanLogURL + userId.toString()
+      axiosDeleteCurrentBanLogURL + userId?.toString() + "/ban-history"
     );
     return response;
   } catch (error) {
@@ -338,41 +270,44 @@ export const axiosDeleteCurrentBanLog = async (userId: number) => {
   }
 };
 
-const axiosGetBrokenCabinetListURL =
-  "/api/admin/search/cabinet/broken/?page=0&length=0";
+const axiosGetBrokenCabinetListURL = "/v4/admin/cabinets/status/BROKEN";
 export const axiosGetBrokenCabinetList = async () => {
   try {
-    const response = await instance.get(axiosGetBrokenCabinetListURL);
+    const response = await instance.get(axiosGetBrokenCabinetListURL, {
+      params: { page: 0, size: 0 },
+    });
     return response.data.result;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosGetBannedUserListURL =
-  "/api/admin/search/user/banned/?page=0&length=0";
+const axiosGetBannedUserListURL = "v4/admin/statistics/users/banned";
 export const axiosGetBannedUserList = async () => {
   try {
-    const response = await instance.get(axiosGetBannedUserListURL);
+    const response = await instance.get(axiosGetBannedUserListURL, {
+      params: { page: 0, size: 0 },
+    });
     return response.data.result;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosGetStatisticsURL = "/api/admin/search/cabinet/statistics/?";
-export const axiosGetStatistics = async (start: number, end: number) => {
+const axiosGetStatisticsURL = "/v4/admin/statistics/lent-histories";
+export const axiosGetStatistics = async (start: Date, end: Date) => {
   try {
-    const response = await instance.get(
-      `${axiosGetStatisticsURL}start=${start}&end=${end}`
-    );
+    const response = await instance.get(axiosGetStatisticsURL, {
+      params: { startDate: start, endDate: end },
+    });
     return response.data;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosGetCabinetNumbersPerFloorURL = "/api/admin/cabinet/count/floor";
+const axiosGetCabinetNumbersPerFloorURL =
+  "/v4/admin/statistics/buildings/floors/cabinets";
 export const axiosGetCabinetNumbersPerFloor = async () => {
   try {
     const response = await instance.get(axiosGetCabinetNumbersPerFloorURL);
@@ -382,27 +317,29 @@ export const axiosGetCabinetNumbersPerFloor = async () => {
   }
 };
 
-const axiosGetOverdueUserListURL =
-  "api/admin/search/user/overdue?page=0&length=0";
+const axiosGetOverdueUserListURL = "/v4/admin/statistics/users/overdue";
 export const axiosGetOverdueUserList = async () => {
   try {
-    const response = await instance.get(axiosGetOverdueUserListURL);
+    const response = await instance.get(axiosGetOverdueUserListURL, {
+      params: { page: 0, size: 0 },
+    });
     return response.data.result;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosGetCabinetLentLogURL = "/api/admin/log/cabinet/";
+const axiosGetCabinetLentLogURL = "/v4/admin/cabinets/";
 export const axiosGetCabinetLentLog = async (
-  cabinetId: number,
+  cabinetId: number | null,
   page: number
 ): Promise<any> => {
+  if (cabinetId == null) return;
   try {
     const response = await instance.get(
-      axiosGetCabinetLentLogURL + cabinetId.toString(),
+      axiosGetCabinetLentLogURL + cabinetId.toString() + "/lent-histories",
       {
-        params: { page: page, length: 10 },
+        params: { page: page, size: 8 },
       }
     );
     return response;
@@ -411,18 +348,98 @@ export const axiosGetCabinetLentLog = async (
   }
 };
 
-const axiosGetUserLentLogURL = "/api/admin/log/user/";
+const axiosGetUserLentLogURL = "/v4/admin/users/";
 export const axiosGetUserLentLog = async (
-  userId: number,
+  userId: number | null,
   page: number
 ): Promise<any> => {
+  if (userId == null) return;
   try {
     const response = await instance.get(
-      axiosGetUserLentLogURL + userId.toString(),
+      axiosGetUserLentLogURL + userId.toString() + "/lent-histories",
       {
-        params: { page: page, length: 10 },
+        params: { page: page, size: 8 },
       }
     );
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosGetClubUserLogURL = "/v4/admin/users/clubs";
+export const axiosGetClubUserLog = async (
+  page: number,
+  size: number
+): Promise<any> => {
+  try {
+    const response = await instance.get(axiosGetClubUserLogURL, {
+      params: { page: page, size: size },
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosCreateClubUserURL = "/v4/admin/users/club";
+export const axiosCreateClubUser = async (
+  clubName: string | null
+): Promise<any> => {
+  if (clubName === null) return;
+  try {
+    const response = await instance.post(axiosCreateClubUserURL, { clubName });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosEditClubUserURL = "/v4/admin/users/club/";
+export const axiosEditClubUser = async (
+  clubInfo: ClubUserDto | null
+): Promise<any> => {
+  if (clubInfo === null || clubInfo.name === null) return;
+  try {
+    const response = await instance.patch(
+      axiosEditClubUserURL + clubInfo.userId.toString(),
+      {
+        clubName: clubInfo.name,
+      }
+    );
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosDeleteClubUserURL = "/v4/admin/users/club/";
+export const axiosDeleteClubUser = async (
+  clubId: number | null
+): Promise<any> => {
+  if (clubId === null) return;
+  try {
+    const response = await instance.delete(
+      axiosDeleteClubUserURL + clubId?.toString()
+    );
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosLentClubUserURL = "/v4/admin/cabinets/club";
+export const axiosLentClubUser = async (
+  userId: number,
+  cabinetId: number,
+  statusNote: string | null
+): Promise<any> => {
+  try {
+    const response = await instance.patch(axiosLentClubUserURL, {
+      userId,
+      cabinetId,
+      statusNote,
+    });
     return response;
   } catch (error) {
     throw error;
